+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+### 4.1.0(41)
+
+VModal
+
+- Modal now supports `iOS` `14.0`
+- Issue with screen-sized modal in landscape is fixed
+
+VBottomSheet
+
+- Issue with screen-sized modal in landscape is fixed
+
+VSideBar
+
+- SideBar now supports `iOS` `14.0`
+- Issue with screen-sized modal in landscape is fixed
+
 ### [4.0.7(40)](https://github.com/VakhoKontridze/VComponents/releases/tag/4.0.7) — *2023 04 20*
 
 General